Further cleanup of unit tests
