+++ ResearchUXFactoryTests/EqualityTests.m
